Copter: rename control_xx.cpp files to mode_xx.cpp
