+++ ArduCopter/mode.h
