+++ ArduCopter/mode_auto.cpp
