+++ ArduCopter/mode_avoid_adsb.cpp
