+++ ArduCopter/mode_flip.cpp
