+++ ArduCopter/mode_guided.cpp
