+++ ArduCopter/mode_guided_nogps.cpp
