+++ ArduCopter/mode_loiter.cpp
