+++ ArduCopter/mode_rtl.cpp
