+++ ArduCopter/mode_smart_rtl.cpp
