+++ ArduCopter/mode_sport.cpp
