+++ ArduCopter/mode_stabilize.cpp
