+++ ArduCopter/mode_throw.cpp
